Fix aliasing violations, optimize
Fix some violations of aliasing rules and optimize the code while at it:
There's no reason to use size_t/ssize_t (typically 64-bit values) for
the size of objects in pixels.

author	Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>

+++ camera/AppCallbackNotifier.cpp
@@ -1061,13 +1061,13 @@ void AppCallbackNotifier::notifyFrame()
                                 void *y_uv[2];
                                 mapper.lock((buffer_handle_t)vBuf, CAMHAL_GRALLOC_USAGE, bounds, y_uv);
 
-                                structConvImage input =  {frame->mWidth,
-                                                          frame->mHeight,
+                                structConvImage input =  {(int)frame->mWidth,
+                                                          (int)frame->mHeight,
                                                           4096,
                                                           IC_FORMAT_YCbCr420_lp,
                                                           (mmByte *)frame->mYuv[0],
                                                           (mmByte *)frame->mYuv[1],
-                                                          frame->mOffset};
+                                                          (int)frame->mOffset};
 
                                 structConvImage output = {mVideoWidth,
                                                           mVideoHeight,
